close #2, changes hashtable to hash map (sorry). I will abide by this rule in the future.
