+++ src/main/java/com/dougyoung/stockssimulator/controller/OrderController.java
@@ -3,22 +3,24 @@ package com.dougyoung.stockssimulator.controller;
 import com.dougyoung.stockssimulator.model.Order;
 import com.dougyoung.stockssimulator.model.Side;
 import com.dougyoung.stockssimulator.model.Stock;
-import com.dougyoung.stockssimulator.service.StockService;
+import com.dougyoung.stockssimulator.service.StockServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/order")
 public class OrderController {
     // TODO: make the controller do literally anything cooler than this.
 
     @Autowired
-    StockService stockService;
+    StockServiceImpl stockServiceImpl;
 
-    @RequestMapping("/order")
-    public Order order(@RequestParam(value="acronym", defaultValue="NIL") String name) {
-        Stock dummyStock = stockService.getStock("AMD");
+    @GetMapping("{acronym}")
+    public Order order(@RequestParam(value="acronym") String name) {
+        Stock dummyStock = stockServiceImpl.getStock("AMD");
         return new Order(dummyStock, Side.BUY, 41.0);
     }
 }
